Revert "Merge branch 'feature-search-taxonomies' into feature/hotel-schema"
This reverts commit 58217e49e308c5b7a6befdc8b681d1fc7e1ddf83, reversing
changes made to 773c104de393b4a367ad8d99213ed6eec634caf3.
