--- schemas/content-fragments/hotels/contact/contact.ts
@@ -1,40 +1,0 @@
-export default {
-  title: "Contact",
-  name: "contact",
-  type: "document",
-  fields: [
-    {
-      name: "title",
-      title: "Title",
-      type: "string",
-    },
-    {
-      name: "phone",
-      title: "Phone",
-      type: "array",
-      of: [
-        {
-          type: "object",
-          fields: [
-            { name: "mobile", type: "string", title: "Mobile" },
-            { name: "type", type: "string", title: "Type" },
-          ],
-        },
-      ],
-    },
-    {
-      name: "email",
-      title: "Email",
-      type: "array",
-      of: [
-        {
-          type: "object",
-          fields: [
-            { name: "email", type: "string", title: "Email" },
-            { name: "type", type: "string", title: "Type" },
-          ],
-        },
-      ],
-    },
-  ],
-};
